mung it all into one file for now, work on makefiles
