--- app/config.py
@@ -1,3 +1,0 @@
-# Per-podcast configs
-from dataclasses import dataclass
-
